+++ src/tg2go/bot/handlers/staff/callbacks/settings.py
@@ -1,15 +1,14 @@
-
 from aiogram import F, Router, types
 from aiogram.fsm.context import FSMContext
 
 from tg2go.bot.handlers.staff.callbacks.category_action.add import AddCategoryStates
 from tg2go.bot.handlers.staff.menus.category import CategoryMenu
+from tg2go.bot.handlers.staff.menus.panel import PanelMenu
 from tg2go.bot.handlers.staff.menus.settings import (
     SettingsAction,
     SettingsCallbackData,
     SettingsCategoryCallbackData,
 )
-from tg2go.bot.handlers.staff.menus.staff import StaffMenu
 from tg2go.bot.lib.message.io import SendMessage
 
 router = Router()
@@ -20,7 +19,6 @@ router = Router()
 )
 async def SettingsAddCategory(
     callback_query: types.CallbackQuery,
-    callback_data: SettingsCallbackData,
     state: FSMContext,
 ) -> None:
     assert isinstance(callback_query.message, types.Message)
@@ -30,10 +28,7 @@ async def SettingsAddCategory(
 
     await SendMessage(
         chat_id=callback_query.message.chat.id,
-        text="Категория нужна для группировки позиций ассортимента."
-        "\nПрисваиваемый категории индекс отвечает за позицию категории относительно других."
-        "\nЧем больше индекс, тем ниже категория в списке."
-        "\n\nПример:\n(имя=A, индекс=5), (имя=B, индекс=1), (имя=C, индекс=7)\nбудут идти как\nB\nA\nC",
+        text="Категория нужна для группировки позиций ассортимента.\nПрисваиваемый категории индекс отвечает за позицию категории относительно других.\nЧем больше индекс, тем ниже категория в списке.\n\nПример:\n(имя=A, индекс=5), (имя=B, индекс=1), (имя=C, индекс=7)\nбудут идти как\nB\nA\nC",
     )
     await SendMessage(
         chat_id=callback_query.message.chat.id,
@@ -50,10 +45,7 @@ async def SettingsCategory(
 ) -> None:
     assert isinstance(callback_query.message, types.Message)
 
-    menu = await CategoryMenu(
-        chat_id=callback_query.message.chat.id,
-        category_id=callback_data.category_id,
-    )
+    menu = await CategoryMenu(callback_data.category_id)
 
     await callback_query.message.edit_text(
         text=menu.text,
@@ -63,13 +55,10 @@ async def SettingsCategory(
 
 
 @router.callback_query(SettingsCallbackData.filter(F.action == SettingsAction.Back))
-async def SettingsBack(
-    callback_query: types.CallbackQuery,
-    callback_data: SettingsCallbackData,
-) -> None:
+async def SettingsBack(callback_query: types.CallbackQuery) -> None:
     assert isinstance(callback_query.message, types.Message)
 
-    menu = StaffMenu(callback_query.message.chat.id)
+    menu = PanelMenu()
 
     await callback_query.message.edit_text(
         text=menu.text,