+++ src/tg2go/bot/handlers/staff/commands/staff.py
@@ -2,7 +2,7 @@ from aiogram import Router, types
 from aiogram.filters.command import Command
 from aiogram.filters.state import StateFilter
 
-from tg2go.bot.handlers.staff.menus.staff import StaffMenu
+from tg2go.bot.handlers.staff.menus.panel import PanelMenu
 from tg2go.bot.lib.message.filters import StaffFilter
 from tg2go.bot.lib.message.io import SendMessage
 
@@ -11,7 +11,7 @@ router = Router()
 
 @router.message(Command("staff"), StateFilter(None), StaffFilter())
 async def CommandStaff(message: types.Message) -> None:
-    menu = StaffMenu(message.chat.id)
+    menu = PanelMenu()
 
     await SendMessage(
         chat_id=message.chat.id,