+++ src/tg2go/bot/handlers/staff/menus/menu.py
@@ -1,9 +1,0 @@
-from dataclasses import dataclass
-
-from aiogram.types import InlineKeyboardMarkup
-
-
-@dataclass
-class Menu:
-    text: str
-    reply_markup: InlineKeyboardMarkup | None