+++ src/tg2go/bot/handlers/staff/menus/settings.py
@@ -1,29 +1,25 @@
-from enum import Enum
-
 from aiogram.filters.callback_data import CallbackData
 from aiogram.types import InlineKeyboardButton, InlineKeyboardMarkup
 
-from tg2go.bot.handlers.staff.menus.menu import Menu
+from tg2go.bot.handlers.staff.menus.common import CreateButton, Menu, StaffAction
 from tg2go.db.models.common.types import CategoryId
 from tg2go.services.staff.category import StaffCategoryService
 
 
-class SettingsAction(str, Enum):
+class SettingsAction(StaffAction):
     AddCategory = "Добавить категорию"
     Back = "Вернуться обратно"
 
 
-class SettingsCallbackData(CallbackData, prefix="s.settings"):
+class SettingsCallbackData(CallbackData, prefix="staff.settings"):
     action: SettingsAction
-    chat_id: int
 
 
-class SettingsCategoryCallbackData(CallbackData, prefix="s.settings.cat"):
+class SettingsCategoryCallbackData(CallbackData, prefix="staff.settings.cat"):
     category_id: CategoryId
-    chat_id: int
 
 
-async def SettingsMenu(chat_id: int) -> Menu:
+async def SettingsMenu() -> Menu:
     srv = StaffCategoryService.Create()
     categories = await srv.GetSortedCategories()
 
@@ -31,32 +27,19 @@ async def SettingsMenu(chat_id: int) -> Menu:
     text = "🔴 Бот не работает\n\nВы находитесь в настройках бота.\n..."
     buttons = [
         [
-            InlineKeyboardButton(
-                text=SettingsAction.AddCategory.value,
-                callback_data=SettingsCallbackData(
-                    action=SettingsAction.AddCategory,
-                    chat_id=chat_id,
-                ).pack(),
-            )
+            CreateButton(cb=SettingsCallbackData, action=SettingsAction.AddCategory),
         ],
         [
             InlineKeyboardButton(
                 text=cat.name,
                 callback_data=SettingsCategoryCallbackData(
-                    category_id=cat.category_id,
-                    chat_id=chat_id,
+                    category_id=cat.category_id
                 ).pack(),
             )
             for cat in categories
         ],
         [
-            InlineKeyboardButton(
-                text=SettingsAction.Back.value,
-                callback_data=SettingsCallbackData(
-                    action=SettingsAction.Back,
-                    chat_id=chat_id,
-                ).pack(),
-            )
+            CreateButton(cb=SettingsCallbackData, action=SettingsAction.Back),
         ],
     ]