+++ src/tg2go/bot/handlers/staff/register.py
@@ -1,6 +1,6 @@
 from aiogram import Dispatcher
 
-from tg2go.bot.handlers.staff.callbacks import category, good, settings, staff as stff
+from tg2go.bot.handlers.staff.callbacks import category, good, panel as stff, settings
 from tg2go.bot.handlers.staff.callbacks.category_action import add as category_add
 from tg2go.bot.handlers.staff.callbacks.good_action import add as good_add
 from tg2go.bot.handlers.staff.commands import staff
@@ -11,10 +11,8 @@ def RegisterStaffHandlers(dp: Dispatcher) -> None:
         staff.router,
         stff.router,
         settings.router,
-
         category.router,
         category_add.router,
-
         good.router,
         good_add.router,
     )