+++ CHANGELOG.md
@@ -28,6 +28,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.1.0] - 2025-06-20 - Vlad
 
+### Added
+
+- /staff command
+
 ### Updated
 
 - Change type of id of order & good from UUID of int